Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/cssom/MediaList-expected.txt


FAIL CSSOM - MediaList interface assert_equals: MediaList toString method expected "screen and (min-width: 480px), print, projection" but got "[object MediaList]"
